requirements.txt
annotated-types==0.6.0
anyio==4.3.0
aspose-3d==23.12.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
pillow == 10.2.0
filelock==3.13.1
fsspec==2023.12.2
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
google-generativeai   ==  0.4.1
llvmlite==0.41.1
MarkupSafe==2.1.5
marshmallow==3.20.1
openai == 1.14.2
more-itertools==10.1.0
mpmath==1.3.0
msvc-runtime==14.34.31931
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
orjson==3.9.15
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pygltflib==1.16.1
pymeshlab==2023.12
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tiktoken==0.5.2
torch==2.1.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.28.0
watchfiles==0.21.0
websocket==0.2.1
websockets==12.0
wrapt==1.16.0
zope.event==5.0
zope.interface==6.1
